requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r ./base.txt

gunicorn  # https://github.com/benoitc/gunicorn
psycopg2 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
Collectfast  # https://github.com/antonagestam/collectfast
django-anymail[mailgun]
django-redis  # https://github.com/jazzband/django-redis


# Genuine versions
# ----------------------------------------------------------
# gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
# psycopg2==2.8.6 --no-binary psycopg2  # https://github.com/psycopg/psycopg2
# Collectfast==2.2.0  # https://github.com/antonagestam/collectfast
# django-anymail[mailgun]==8.0
# django-redis==4.12.1  # https://github.com/jazzband/django-redis
